Demo

+++ dataproduct.api/Models/NM_Models/STD_NXT_Filter.cs
@@ -6,16 +6,24 @@ namespace dataproduct.api.ResponseModels
     /// <summary>
     /// Model hứng dữ liệu tổng hợp phụ liệu theo ngày/ca từ stored procedure.
     /// </summary>
+    //[Keyless]
+    //public class STD_NXT_Filter
+    //{
+    //    public string? BieuMau { get; set; }
+    //    public decimal Scope { get; set; }
+    //    public decimal ID_PhuLieu { get; set; }
+    //    public string? TenPhuLieu { get; set; }
+    //    public double? TotalKLPhuGia { get; set; }
+    //}
     [Keyless]
     public class STD_NXT_Filter
     {
         public string? BieuMau { get; set; }
-        public decimal Scope { get; set; }
-        public decimal ID_PhuLieu { get; set; }
+        public int Scope { get; set; }
+        public int ID_PhuLieu { get; set; }
         public string? TenPhuLieu { get; set; }
         public double? TotalKLPhuGia { get; set; }
     }
-
     [Keyless]
     public class STD_NXT_Filter_Init
     {
